backend/requirements.txt
asgiref==3.8.1
django==5.1.7
django-cors-headers==4.7.0
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
pip==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
types-pyyaml==6.0.12.20241230
typing-extensions==4.12.2
